update(testingaccessibility): extract cert utils
